Remove custom server
