+++ package.json
@@ -10,13 +10,13 @@
   ],
   "scripts": {
     "analyze": "ANALYZE=true yarn build",
-    "dev": "node server.js",
+    "dev": "next dev",
     "prebuild": "rimraf .next",
     "build": "NODE_ENV=production next build",
     "deploy": "now -t $NOW_TOKEN -T styledcomponents && now alias -t $NOW_TOKEN -T styledcomponents",
     "lint:fix": "eslint . --fix",
     "lint": "eslint .",
-    "start": "NODE_ENV=production node server.js",
+    "start": "NODE_ENV=production next start",
     "test": "jest -c .jest.config.js",
     "precommit": "lint-staged",
     "prettier": "prettier **/*.js --write",
@@ -34,10 +34,8 @@
     "babel-plugin-module-resolver": "^3.1.1",
     "babel-plugin-polished": "^1.1.0",
     "docsearch.js": "2.6.3",
-    "express": "^4.16.3",
     "invariant": "^2.2.4",
     "isomorphic-fetch": "^2.2.1",
-    "lru-cache": "^5.1.1",
     "markdown-to-jsx": "^6.10.2",
     "next": "^9.0.6",
     "polished": "^3.4.0",
@@ -70,7 +68,7 @@
     "husky": "^2.4.0",
     "jest": "^24.8.0",
     "jest-styled-components": "beta",
-    "lint-staged": "^8.2.0",
+    "lint-staged": "^9.3.0",
     "now": "^15.4.0",
     "prettier": "^1.18.2",
     "react-test-renderer": "^16.5.2",
