+++ sections/homepage/getting-started.md
@@ -6,13 +6,13 @@ export default ({ children }) => (
   <Content data-e2e-id="content">
     <AlignCenter>
       <a href="https://github.com/styled-components/styled-components">
-        <Badge src="/proxy/stars.svg" alt="Stars on GitHub" />
+        <Badge src="/api/proxy/stars.svg" alt="Stars on GitHub" />
       </a>
       <a href="https://www.npmjs.com/package/styled-components">
-        <Badge src="/proxy/npm-v.svg" alt="Current version" />
+        <Badge src="/api/proxy/npm-v.svg" alt="Current version" />
       </a>
-      <Badge src="/proxy/downloads.svg" alt="Monthly downloads" />
-      <Badge src="/proxy/size.svg" alt="Gzipped size" />
+      <Badge src="/api/proxy/downloads.svg" alt="Monthly downloads" />
+      <Badge src="/api/proxy/size.svg" alt="Gzipped size" />
       <a href="https://spectrum.chat/styled-components">
         <Badge alt="Join the community on Spectrum" src="https://withspectrum.github.io/badge/badge.svg" />
       </a>
